requirements.txt
gym==0.10.9
numpy==1.15.4
ray==0.6.6
tensorflow-probability==0.7.0
opencv-python==4.1.0.25

tensorboard==1.14.0
tensorboardX==1.6
tensorflow-gpu==1.14.0


